WIP - move xdg_shell and wl_shell specific code to their own modules
I believe this is mostly implemented however I'm not confident this is
the nicest design. I'm unsure how to best modify the `Handler` pattern
to handle the different shells. I've changed the Handler::configure
method to take a `Configure` enum, which provides the shell-specific
data.

The example has not yet been updated as I'm unsure how to check whether
`xdg_shell` is available and fall back to `wl_shell` otherwise. It seems
this process is currently wrapped up by the `wayland_env` macro.